fix agg nullable
