+++ fe/fe-core/src/main/java/org/apache/doris/nereids/rules/RuleType.java
@@ -100,10 +100,6 @@ public enum RuleType {
     ANALYZE_CTE(RuleTypeClass.REWRITE),
     RELATION_AUTHENTICATION(RuleTypeClass.VALIDATION),
 
-    ADJUST_NULLABLE_FOR_PROJECT_SLOT(RuleTypeClass.REWRITE),
-    ADJUST_NULLABLE_FOR_AGGREGATE_SLOT(RuleTypeClass.REWRITE),
-    ADJUST_NULLABLE_FOR_HAVING_SLOT(RuleTypeClass.REWRITE),
-    ADJUST_NULLABLE_FOR_REPEAT_SLOT(RuleTypeClass.REWRITE),
     ADD_DEFAULT_LIMIT(RuleTypeClass.REWRITE),
 
     CHECK_ROW_POLICY(RuleTypeClass.REWRITE),
