+++ regression-test/data/nereids_rules_p0/adjust_nullable/test_agg_nullable.out
@@ -1,0 +1,11 @@
+-- This file is automatically generated. You should know what you did if you want to edit this
+-- !agg_nullable --
+\N
+
+-- !agg_nullable_shape --
+PhysicalResultSink
+--PhysicalProject[AND[k IS NULL,NULL] AS `k > 10 and k < 5`]
+----hashAgg[GLOBAL]
+------hashAgg[LOCAL]
+--------PhysicalEmptyRelation
+
